Prepare the directory structure:   - Create a components directory in the src folder.   - Move App.js, add Display, ButtonPanel and Button component to the directory.
