+++ src/components/App.js
@@ -1,4 +1,4 @@
-import './App.css';
+import '../App.css';
 
 const App = () => (
   <div className="App">
